Simplify DeepSeek V4 checkpoint layout handling
